Add jekyll and styling/layout for future content
Jekyll structure has been added to the project.

A second css file has been created for future pages and applied to the newly created "default" layout. The home page is exempt from both. Reorganizing of the nav on the home page has been done, and the stripe on the right side is now intentional.